Add EmbeddedQueryResultCache
Most of the job is already done by the `EmbeddedQueryDependencyLinksStore`
to track and update query dependencies.

To eliminate necessary SQL/SPARQL connections `EmbeddedQueryResultCache` ought to
cache a subject list returned from the `QueryEngine`.

We don't cache the string result of a printer == it means we don't
interfere with how the printer manipulates the data, `EmbeddedQueryResultCache` as
the name suggests caches the return object (aka `QueryResult`) from
the QueryEngine before it is forwarded to an individual printer.

If for some reason `EmbeddedQueryDependencyLinksStore` is not enabled then
auto-invalidation of the `EmbeddedQueryResultCache` can not occur but instead
(same as of now == manual intervention using the purge button) setting
`$GLOBALS['smwgEmbeddedQueryResultCacheRefreshOnPurge'] = true;` can
be used to emit a purge action event to invalidate all query cache items stored
with the corresponding article (aka subject).

`$GLOBALS['smwgEmbeddedQueryResultCacheLifetime'] = 60 * 60 * 24; // a day` declares the lifetime
of a cached item